node/requirements.txt
numpy
python-bitcoinrpc
